pyproject.toml
[project]
name = "rigol-dho824-mcp"
version = "0.2.0"
description = "Professional MCP server for Rigol DHO800 oscilloscopes, optimized for Sony TC-FX44 tape deck calibration and audio analysis."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "csibug" }
]
keywords = [
    "rigol",
    "oscilloscope",
    "dho824",
    "dho814",
    "dho804",
    "mcp",
    "mcp-server",
    "scpi",
    "test-equipment",
    "measurement",
    "tape-repair",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Hardware",
]
dependencies = [
    "fastmcp>=0.2.0",
    "pyvisa>=1.13.0",
    "pyvisa-py>=0.7.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/csibug/rigol-dho824-mcp"
Repository = "https://github.com/csibug/rigol-dho824-mcp"
Issues = "https://github.com/csibug/rigol-dho824-mcp/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rigol_dho824_mcp"]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.ruff]
line-length = 120
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "build",
    "dist",
    "venv",
    "data",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]